Check if directory exists before trying to remove it.

+++ pyiron/base/job/core.py
@@ -414,10 +414,8 @@ class JobCore(PyironObject):
             if os.path.isfile(self.project_hdf5.file_name):
                 os.remove(self.project_hdf5.file_name)
                 dir_name = self.project_hdf5.file_name.split('.h5')[0] + '_hdf5'
-                try:
+                if os.path.isdir(dir_name):
                     os.rmdir(dir_name)
-                except OSError:
-                    print('Library jobs have no job folder {}'.format(dir_name))
         if self.job_id:
             self.project.db.delete_item(self.job_id)
 
@@ -802,6 +800,12 @@ class JobCore(PyironObject):
             pass  # no name check in Python 2.7
 
     def compress(self, files_to_compress=None):
+        """
+        Compress the output files of a job object.
+
+        Args:
+            files_to_compress (list):
+        """
         if not any([".tar.bz2" in file for file in self.list_files()]):
             if files_to_compress is None:
                 files_to_compress = list(self.list_files())
@@ -814,7 +818,7 @@ class JobCore(PyironObject):
                             tar.add(name)
                 for name in files_to_compress:
                     if "tar" not in name:
-                        fullname=os.path.join(self.working_directory, name)
+                        fullname = os.path.join(self.working_directory, name)
                         if os.path.isfile(fullname):
                             os.remove(fullname)
                         elif os.path.isdir(fullname):
@@ -825,8 +829,11 @@ class JobCore(PyironObject):
             print('The files are already compressed!')
 
     def decompress(self):
+        """
+        Decompress the output files of a compressed job object.
+        """
         try:
-            tar_file_name=os.path.join(self.working_directory, self.job_name + ".tar.bz2")
+            tar_file_name = os.path.join(self.working_directory, self.job_name + ".tar.bz2")
             with tarfile.open(tar_file_name, "r:bz2") as tar:
                 tar.extractall(self.working_directory)
             os.remove(tar_file_name)
@@ -834,6 +841,12 @@ class JobCore(PyironObject):
             pass
 
     def is_compressed(self):
+        """
+        Check if the job is already compressed or not.
+
+        Returns:
+            bool: [True/False]
+        """
         compressed_name = self.job_name + ".tar.bz2"
         for name in self.list_files():
             if compressed_name in name:
@@ -844,21 +857,21 @@ class JobCore(PyironObject):
         fpath = self.project_hdf5.file_path
         jname = self.job_name
         h5_dir_name = jname + "_hdf5"
-        h5_file_name =jname + ".h5"
+        h5_file_name = jname + ".h5"
         # assert os.path.isdir(h5_dir_name)
         # assert os.path.isfile(h5_file_name)
         try:
             cwd = os.getcwd()
             os.chdir(fpath)
             with tarfile.open(os.path.join(fpath, self.job_name + ".tar.bz2"), "w:bz2") as tar:
-                for name in [h5_dir_name,h5_file_name]:
-                        tar.add(name)
-            for name in [h5_dir_name,h5_file_name]:
-                     fullname = os.path.join(fpath, name)
-                     if os.path.isfile(fullname):
-                         os.remove(fullname)
-                     elif os.path.isdir(fullname):
-                         shutil.rmtree(fullname)
+                for name in [h5_dir_name, h5_file_name]:
+                    tar.add(name)
+            for name in [h5_dir_name, h5_file_name]:
+                fullname = os.path.join(fpath, name)
+                if os.path.isfile(fullname):
+                    os.remove(fullname)
+                elif os.path.isdir(fullname):
+                    shutil.rmtree(fullname)
         finally:
             os.chdir(cwd)
